refactor: move ActResourceServerSpec out of internal/, matching existing convention
ActResourceServerSpec is part of the public API (re-exported from
index.ts), but it was defined under internal/, which is reserved for
implementation details never meant to be imported by consumers.
Existing public types (ActJobExecResult, ActExecStatus, ...) all live
as top-level modules directly under src/, with internal/ code
importing them by reaching up (e.g. ActExecListener.ts imports
ActJobExecResult from '../ActJobExecResult.js').

Moved ActResourceServerSpec to src/ActResourceServerSpec.ts to match
that convention. This keeps the type import-cycle-free exactly as
before (the module still has no dependencies of its own), while fixing
the misleading location.
